requirements.txt
# Web 框架
flask==3.1.0
werkzeug==3.1.3

# 生产服务器
gunicorn==23.0.0

# 腾讯云 SDK
tencentcloud-sdk-python==3.0.1290

# 安全增强
flask-limiter==3.8.0  # 速率限制

# 开发和测试工具 (可选)
# pytest==8.3.4
# safety==3.2.11
# bandit==1.8.0
